fix lattice rescoring, lattice read/write, node score
git-svn-id: svn+ssh://svn.code.sf.net/p/cmusphinx/code/trunk/sphinx4@12801 94700074-3cef-4d97-a70e-9c8c206c02f5

+++ sphinx4-core/src/main/java/edu/cmu/sphinx/result/Edge.java
@@ -59,7 +59,7 @@ public class Edge {
 
         String from = tokens.nextToken();
         String to = tokens.nextToken();
-        int score = Integer.parseInt(tokens.nextToken());
+        double score = Double.parseDouble(tokens.nextToken());
 
         Node fromNode = lattice.getNode(from);
         if (fromNode == null) {
